# swagger/center.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OKR 系统API",
    "description": "OKR 系统相关接口文档！",
    "contact": {
      "name": "发际线与我作对",
      "url": "https://github.com/CarefreeState/OKR-System",
      "email": "2040484356@qq.com"
    },
    "version": "3.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:1701",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
...
    "/team/grant": {
      "post": {
        "tags": [
          "团队 OKR 测试接口"
        ],
        "summary": "给成员授权，使其可以扩展一个子团队",
        "operationId": "grantTeamForMemberUsingPOST",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/授权成员所需数据"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/team/list": {
      "get": {
        "tags": [
          "团队 OKR 测试接口"
        ],
        "summary": "获取团队 OKR 列表",
        "operationId": "getTeamOkrsUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«List«团队 OKR 部分数据»»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/team/qrcode/{teamId}": {
      "post": {
        "tags": [
          "团队 OKR 测试接口"
        ],
        "summary": "获取邀请码",
        "operationId": "getQRCodeUsingPOST",
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "团队 OKR ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«string»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/team/tree/child/{id}": {
      "post": {
        "tags": [
          "团队 OKR 测试接口"
        ],
        "summary": "获取一个团队的子树",
        "operationId": "getChildTreeUsingPOST",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "团队 OKR ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«List«团队 OKR 统计数据»»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/team/tree/{id}": {
      "post": {
        "tags": [
          "团队 OKR 测试接口"
        ],
        "summary": "获取一个团队所在的树",
        "operationId": "getCompleteTreeUsingPOST",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "团队 OKR ID",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«List«团队 OKR 统计数据»»"
                }
              }
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/teampersonal/list": {
      "get": {
        "tags": [
          "团队个人 OKR 测试接口"
        ],
        "summary": "获取团队个人 OKR 列表",
        "operationId": "getTeamOkrsUsingGET_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/统一响应«List«团队个人 OKR 部分数据»»"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "授权成员所需数据": {
        "title": "授权成员所需数据",
        "type": "object",
        "properties": {
          "teamId": {
            "type": "integer",
            "description": "团队 OKR ID",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "用户 ID",
            "format": "int64"
          }
        }
      },
...
      "统一响应": {
        "title": "统一响应",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "具体数据"
          },
          "message": {
            "type": "string",
            "description": "描述语"
          }
        }
      },
...
      "统一响应«string»": {
        "title": "统一响应«string»",
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码",
            "format": "int32"
          },
          "data": {
            "type": "string",
            "description": "具体数据"
          },
          "message": {
            "type": "string",
            "description": "描述语"
          }
        }
      },
...
    }
  }
}